Refactored completeness pdf section into separate component
